Make blkreftable API use size_t/ssize_t consistently
It was using int for the input and output length, where size_t or
ssize_t would be more appropriate.  It might not matter in practice,
but it makes the APIs consistent at all levels.

Discussion: https://www.postgresql.org/message-id/flat/f9aab072-0078-49e4-ab93-3b08086a4406@eisentraut.org

+++ src/backend/backup/walsummary.c
@@ -269,11 +269,11 @@ IsWalSummaryFilename(char *filename)
 /*
  * Data read callback for use with CreateBlockRefTableReader.
  */
-int
-ReadWalSummary(void *wal_summary_io, void *data, int length)
+size_t
+ReadWalSummary(void *wal_summary_io, void *data, size_t length)
 {
 	WalSummaryIO *io = wal_summary_io;
-	int			nbytes;
+	ssize_t		nbytes;
 
 	nbytes = FileRead(io->file, data, length, io->filepos,
 					  WAIT_EVENT_WAL_SUMMARY_READ);
@@ -290,11 +290,11 @@ ReadWalSummary(void *wal_summary_io, void *data, int length)
 /*
  * Data write callback for use with WriteBlockRefTable.
  */
-int
-WriteWalSummary(void *wal_summary_io, void *data, int length)
+size_t
+WriteWalSummary(void *wal_summary_io, void *data, size_t length)
 {
 	WalSummaryIO *io = wal_summary_io;
-	int			nbytes;
+	ssize_t		nbytes;
 
 	nbytes = FileWrite(io->file, data, length, io->filepos,
 					   WAIT_EVENT_WAL_SUMMARY_WRITE);
@@ -306,7 +306,7 @@ WriteWalSummary(void *wal_summary_io, void *data, int length)
 	if (nbytes != length)
 		ereport(ERROR,
 				(errcode_for_file_access(),
-				 errmsg("could not write file \"%s\": wrote only %d of %d bytes at offset %lld",
+				 errmsg("could not write file \"%s\": wrote only %zd of %zu bytes at offset %lld",
 						FilePathName(io->file), nbytes,
 						length, (long long) io->filepos),
 				 errhint("Check free disk space.")));
